goups screen -> subGroups
change between horizontal list to verical list
